style: splitting notify token into files
