+++ Assets/Mirage/Runtime/SocketLayer/Connection/INotifyCallBack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 968b5fcba723a2844902db2f6c55443e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
